Update Region model and migrations to include imagepath field

+++ database/migrations/2024_03_14_000411_create_invoices_table.php
@@ -18,10 +18,10 @@ return new class extends Migration
             $table->date('date');
             $table->date('issue_date');
             $table->date('due_date');
-            $table->decimal('consumption', 8, 2);
+            $table->integer('consumption');
             $table->date('nextIndexReading')->nullable();
             $table->date('pleasePayBefore')->nullable();
-            $table->decimal('amount', 8, 2);
+            $table->decimal('amount', 8, 3);
             $table->enum('payment_status', Invoice::$PAYMENT_STATUSES);
             $table->enum('period', Invoice::$PERIODS);
             $table->foreignId('local_id')->constrained();
